+++ CHANGELOG.md
@@ -16,6 +16,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com),
 and this project adheres to [Semantic Versioning](https://semver.org).
 
 ## [Unreleased]
+### Added
+- added support for properties `-r` and `-X` for remote test execution (#50)
 
 ## [2.3.0]
 ### Added